displaying list of registered users
